requirements.txt
kafka-python
Pillow==9.5.0
easyocr
numpy